Move back onNewToken to Firebase service.
