+++ android/src/main/java/com/dieam/reactnativepushnotification/modules/RNPushNotificationListenerService.java
@@ -3,7 +3,20 @@ package com.dieam.reactnativepushnotification.modules;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 
+import android.os.Handler;
+import android.os.Looper;
+import android.util.Log;
+import android.support.annotation.NonNull;
+
 import com.dieam.reactnativepushnotification.modules.RNReceivedMessageHandler;
+import com.facebook.react.ReactApplication;
+import com.facebook.react.ReactInstanceManager;
+import com.facebook.react.bridge.Arguments;
+import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReactContext;
+import com.facebook.react.bridge.WritableMap;
+
+import static com.dieam.reactnativepushnotification.modules.RNPushNotification.LOG_TAG;
 
 public class RNPushNotificationListenerService extends FirebaseMessagingService {
 
@@ -11,7 +24,40 @@ public class RNPushNotificationListenerService extends FirebaseMessagingService 
 
     @Override
     public void onNewToken(String token) {
-        mMessageReceivedHandler.onNewToken(token);
+        final String deviceToken = token;
+        Log.d(LOG_TAG, "Refreshed token: " + deviceToken);
+
+        Handler handler = new Handler(Looper.getMainLooper());
+        handler.post(new Runnable() {
+            public void run() {
+                // Construct and load our normal React JS code bundle
+                ReactInstanceManager mReactInstanceManager = ((ReactApplication) getApplication()).getReactNativeHost().getReactInstanceManager();
+                ReactContext context = mReactInstanceManager.getCurrentReactContext();
+                // If it's constructed, send a notification
+                if (context != null) {
+                    handleNewToken((ReactApplicationContext) context, deviceToken);
+                } else {
+                    // Otherwise wait for construction, then send the notification
+                    mReactInstanceManager.addReactInstanceEventListener(new ReactInstanceManager.ReactInstanceEventListener() {
+                        public void onReactContextInitialized(ReactContext context) {
+                            handleNewToken((ReactApplicationContext) context, deviceToken);
+                        }
+                    });
+                    if (!mReactInstanceManager.hasStartedCreatingInitialContext()) {
+                        // Construct it in the background
+                        mReactInstanceManager.createReactContextInBackground();
+                    }
+                }
+            }
+        });
+    }
+
+    private void handleNewToken(ReactApplicationContext context, String token) {
+        RNPushNotificationJsDelivery jsDelivery = new RNPushNotificationJsDelivery(context);
+
+        WritableMap params = Arguments.createMap();
+        params.putString("deviceToken", token);
+        jsDelivery.sendEvent("remoteNotificationsRegistered", params);
     }
 
     @Override
