+++ android/src/main/java/com/dieam/reactnativepushnotification/modules/RNReceivedMessageHandler.java
@@ -1,6 +1,5 @@
 package com.dieam.reactnativepushnotification.modules;
 
-import java.util.Map;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
 
@@ -23,56 +22,18 @@ import com.facebook.react.bridge.WritableMap;
 
 import org.json.JSONObject;
 
+import java.util.Map;
 import java.util.List;
 import java.util.Random;
 
 import static android.content.Context.ACTIVITY_SERVICE;
 import static com.dieam.reactnativepushnotification.modules.RNPushNotification.LOG_TAG;
 
-
-
 public class RNReceivedMessageHandler {
     private FirebaseMessagingService mFirebaseMessagingService;
 
     public RNReceivedMessageHandler(@NonNull FirebaseMessagingService service) {
         this.mFirebaseMessagingService = service;
-    }
-
-    public void onNewToken(String token) {
-        final String deviceToken = token;
-        Log.d(LOG_TAG, "Refreshed token: " + deviceToken);
-
-        Handler handler = new Handler(Looper.getMainLooper());
-        handler.post(new Runnable() {
-            public void run() {
-                // Construct and load our normal React JS code bundle
-                ReactInstanceManager mReactInstanceManager = ((ReactApplication) mFirebaseMessagingService.getApplication()).getReactNativeHost().getReactInstanceManager();
-                ReactContext context = mReactInstanceManager.getCurrentReactContext();
-                // If it's constructed, send a notification
-                if (context != null) {
-                    handleNewToken((ReactApplicationContext) context, deviceToken);
-                } else {
-                    // Otherwise wait for construction, then send the notification
-                    mReactInstanceManager.addReactInstanceEventListener(new ReactInstanceManager.ReactInstanceEventListener() {
-                        public void onReactContextInitialized(ReactContext context) {
-                            handleNewToken((ReactApplicationContext) context, deviceToken);
-                        }
-                    });
-                    if (!mReactInstanceManager.hasStartedCreatingInitialContext()) {
-                        // Construct it in the background
-                        mReactInstanceManager.createReactContextInBackground();
-                    }
-                }
-            }
-        });
-    }
-
-    private void handleNewToken(ReactApplicationContext context, String token) {
-        RNPushNotificationJsDelivery jsDelivery = new RNPushNotificationJsDelivery(context);
-
-        WritableMap params = Arguments.createMap();
-        params.putString("deviceToken", token);
-        jsDelivery.sendEvent("remoteNotificationsRegistered", params);
     }
 
     public void handleReceivedMessage(RemoteMessage message) {
